backend/requirements.txt
psycopg[binary]==3.2.13
psycopg-pool>=3.2.4
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
Pillow==11.3.0
